+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Cleanup:
   * Remove `p_messageOutput` from Context of `uctBtnExport`, as Parameter `UCT_ExportUserLog` does not exist Anymore
   * Remove Outdated TODOs
+* Quality: Add testcase for swap without message-output
 * Dokumentation: Fix Einrückung in Readme 
 * Chore: Replace Deprecated GitHub Actions